Refactored Pillar2Error
