+++ app/uk/gov/hmrc/pillar2/controllers/actions/AuthAction.scala
@@ -20,7 +20,7 @@ import com.google.inject.{ImplementedBy, Inject}
 import play.api.mvc.*
 import uk.gov.hmrc.auth.core.{AuthConnector, AuthorisationException, AuthorisedFunctions}
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.pillar2.models.errors.AuthorizationError
+import uk.gov.hmrc.pillar2.models.errors.Pillar2Error.AuthorizationError
 import uk.gov.hmrc.play.http.HeaderCarrierConverter
 
 import scala.concurrent.{ExecutionContext, Future}
