+++ app/uk/gov/hmrc/pillar2/models/errors/Pillar2Error.scala
@@ -17,38 +17,42 @@
 package uk.gov.hmrc.pillar2.models.errors
 
 sealed trait Pillar2Error extends Exception {
-  val message: String
-  val code:    String
+  def message: String
+  def code:    String
 
   override def getMessage: String = s"Code: '$code' Message: '$message'"
 }
 
-case class MissingHeaderError(headerName: String) extends Pillar2Error {
-  val message: String = s"Missing $headerName header"
-  val code:    String = "001"
-}
+object Pillar2Error {
 
-case class InvalidJsonError(decodeError: String) extends Pillar2Error {
-  val message: String = s"Invalid JSON payload: $decodeError"
-  val code:    String = "002"
-}
+  case class MissingHeaderError(headerName: String) extends Pillar2Error {
+    val message: String = s"Missing $headerName header"
+    val code:    String = "001"
+  }
 
-case object ApiInternalServerError extends Pillar2Error {
-  val message: String = "Internal server error"
-  val code:    String = "003"
-}
+  case class InvalidJsonError(decodeError: String) extends Pillar2Error {
+    val message: String = s"Invalid JSON payload: $decodeError"
+    val code:    String = "002"
+  }
 
-case object AuthorizationError extends Pillar2Error {
-  val message: String = "Not Authorized"
-  val code:    String = "401"
-}
+  case object ApiInternalServerError extends Pillar2Error {
+    val message: String = "Internal server error"
+    val code:    String = "003"
+  }
 
-case object SubscriptionProcessingError extends Pillar2Error {
-  val message: String = "Subscription is being processed"
-  val code:    String = "422"
-}
+  case object AuthorizationError extends Pillar2Error {
+    val message: String = "Not Authorized"
+    val code:    String = "401"
+  }
 
-case class ETMPValidationError(validationErrorCode: String, validationError: String) extends Pillar2Error {
-  val message: String = validationError
-  val code:    String = validationErrorCode
+  case object SubscriptionProcessingError extends Pillar2Error {
+    val message: String = "Subscription is being processed"
+    val code:    String = "422"
+  }
+
+  final case class ETMPValidationError(validationErrorCode: String, validationError: String) extends Pillar2Error {
+    val message: String = validationError
+    val code:    String = validationErrorCode
+  }
+
 }
