+++ app/uk/gov/hmrc/pillar2/service/SubscriptionService.scala
@@ -26,7 +26,7 @@ import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2.connectors.SubscriptionConnector
 import uk.gov.hmrc.pillar2.models.*
 import uk.gov.hmrc.pillar2.models.audit.AuditResponseReceived
-import uk.gov.hmrc.pillar2.models.errors.SubscriptionProcessingError
+import uk.gov.hmrc.pillar2.models.errors.Pillar2Error.SubscriptionProcessingError
 import uk.gov.hmrc.pillar2.models.grs.EntityType
 import uk.gov.hmrc.pillar2.models.hods.subscription.common.*
 import uk.gov.hmrc.pillar2.models.hods.subscription.request.RequestDetail
@@ -40,6 +40,7 @@ import uk.gov.hmrc.play.audit.http.connector.AuditResult
 import java.time.LocalDate
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
+
 class SubscriptionService @Inject() (
   repository:            ReadSubscriptionCacheRepository,
   subscriptionConnector: SubscriptionConnector,
