+++ test/uk/gov/hmrc/pillar2/controllers/actions/AuthActionSpec.scala
@@ -33,7 +33,7 @@ import uk.gov.hmrc.auth.core.authorise.Predicate
 import uk.gov.hmrc.auth.core.retrieve.Retrieval
 import uk.gov.hmrc.auth.core.{AuthConnector, MissingBearerToken}
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.pillar2.models.errors.AuthorizationError
+import uk.gov.hmrc.pillar2.models.errors.Pillar2Error.AuthorizationError
 
 import scala.concurrent.duration.*
 import scala.concurrent.{ExecutionContext, Future}
